+++ src/SLCore.UnitTests/SLCoreRpcFactoryTests.cs
@@ -19,6 +19,7 @@
  */
 
 using NSubstitute;
+using SonarLint.VisualStudio.SLCore.Configuration;
 using SonarLint.VisualStudio.SLCore.Core;
 using SonarLint.VisualStudio.SLCore.Core.Process;
