+++ src/SLCore/Core/Process/ISLCoreProcessFactory.cs
@@ -18,6 +18,8 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
+using SonarLint.VisualStudio.SLCore.Configuration;
+
 namespace SonarLint.VisualStudio.SLCore.Core.Process;
 
 internal interface ISLCoreProcessFactory
